Added HomeScreenUI to main
