+++ Index.java
@@ -1,25 +1,10 @@
-import UIComp.LoginScreenUI;
+import UIComp.*;
 import URLHandler.*; 
 
 import java.awt.*;
 import javax.swing.*; 
 import java.awt.event.*;
 
-class Screen extends JFrame{
-	public String id; 
-	Screen(String sc){
-		setTitle(sc); 
-		id = sc;  
-		setSize(1000,600); 
-	  	setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);	
-		setVisible(false); 
-	}
-	
-	public void isVisible(boolean val){
-		setVisible(val); 
-	}
-
-} 
 
 class Users{
 	public String username, email, password;
@@ -50,24 +35,24 @@ public class Index extends Thread{
 	}
 
 	
-	static class HomeScreen extends Screen implements ActionListener{
-		public HomeScreen(String a){
-			super(a);
-			JButton b = new JButton("Switch"); 
-			b.addActionListener(this); 
-			add(b);
+	static class HomeScreen extends HomeScreenUI implements ActionListener{
+		public HomeScreen(){
+			super();
+			logout.addActionListener(this);
 		}
  
+		public static String id = "HomeScreen"; 
 
 		public void actionPerformed(ActionEvent e){
-			screenSetState(ls.id);  
+			user.email = null;
+			user.password = null;
+
+			screenSetState(ls.id);
 		}
 	
 	}
 
-	
-
-	static HomeScreen hs = new HomeScreen("HomeScreen");  
+	static HomeScreen hs = new HomeScreen();  
 	static LoginScreen ls = new LoginScreen();
 	
 	public static void screenSetState(String val){
